Revert "fix(接口测试): 接口测试报告中未执行的请求也要统计到总数中 (#11953)"
This reverts commit 2c819e5a7935a78cc16e4ae9a87d8b5cc7ddf3fe.

+++ backend/src/main/java/io/metersphere/api/dto/definition/request/controller/loop/CountController.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CountController {
     private int loops;
-    private long interval;
+    private int interval;
     private boolean proceed;
     private Object requestResult;
